requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.3.0=mkl
absl-py=0.7.1=pypi_0
astor=0.8.0=pypi_0
atari-py=0.2.3=pypi_0
baselines=0.1.6=dev_0
blas=1.0=mkl
c-ares=1.15.0=h7b6447c_1
ca-certificates=2019.5.15=0
certifi=2019.6.16=py37_0
cffi=1.12.3=py37h2e261b9_0
click=7.0=pypi_0
cloudpickle=1.2.1=pypi_0
cudatoolkit=10.0.130=0
cycler=0.10.0=pypi_0
decorator=4.4.0=pypi_0
freetype=2.9.1=h8a8886c_1
future=0.17.1=pypi_0
gast=0.2.2=pypi_0
google-pasta=0.1.7=py_0
grpcio=1.22.0=pypi_0
gym=0.13.0=pypi_0
h5py=2.9.0=py37h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
imageio=2.5.0=pypi_0
intel-openmp=2019.4=243
joblib=0.13.2=pypi_0
jpeg=9b=h024ee3a_2
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.8.0=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
markdown=3.1.1=py37_0
matplotlib=3.1.1=pypi_0
mkl=2019.4=243
mkl_fft=1.0.12=py37ha843d7b_0
mkl_random=1.0.2=py37hd81dba3_0
ncurses=6.1=he6710b0_1
networkx=2.3=pypi_0
ninja=1.9.0=py37hfd86e86_0
numpy=1.16.4=py37h7e9f1db_0
numpy-base=1.16.4=py37hde5b4d6_0
olefile=0.46=py37_0
opencv-python=4.1.0.25=pypi_0
openssl=1.1.1c=h7b6447c_1
pillow=6.0.0=py37h34e0f95_0
pip=19.1.1=py37_0
protobuf=3.8.0=pypi_0
pybullet=2.5.1=pypi_0
pycparser=2.19=py37_0
pyglet=1.3.2=pypi_0
pyparsing=2.4.0=pypi_0
python=3.7.3=h0371630_0
python-dateutil=2.8.0=pypi_0
pytorch=1.1.0=py3.7_cuda10.0.130_cudnn7.5.1_0
pywavelets=1.0.3=pypi_0
readline=7.0=h7b6447c_5
scikit-image=0.15.0=pypi_0
scipy=1.3.0=pypi_0
setuptools=41.0.1=py37_0
six=1.12.0=py37_0
sqlite=3.28.0=h7b6447c_0
tensorboard=1.14.0=py37hf484d3e_0
tensorflow=1.14.0=mkl_py37h45c423b_0
tensorflow-base=1.14.0=mkl_py37h7ce6ba3_0
tensorflow-estimator=1.14.0=py_0
tensorflow-gpu=1.14.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.8=hbc83047_0
torchvision=0.3.0=py37_cu10.0.130_1
tqdm=4.32.2=pypi_0
visualpriors=0.3=pypi_0
werkzeug=0.15.4=py_0
wheel=0.33.4=py37_0
wrapt=1.11.2=py37h7b6447c_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
